MM-23503: Fix race in websocket_client writer
We create a separate writer goroutine where all the writes happen.
And we also clean up a lot of things related to the reader goroutine.

Now all cleanup related activities happen inside the Close method.

We also fix a bug where pingWatchdog would not have been initialized
if ConnectWithDialer was called after initializing a WebSocketClient
struct manually.

The code has been decorated with a lot of comments to explain the flow
of events. Overall, there are still a lot of edge conditions
which cannot be removed due to compatibility reasons. This entire
client is due for an overhaul in v6.